pyproject.toml
[project]
name = "lesson-automation-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.97.0",
    "openai>=2.33.0",
    "python-dotenv>=1.2.2",
]
